avoid API break : restore original photos.py, rename new implementation photos2.py
